DAQ HAT: Super-Resolution, USB batching & FW UI
Adds Super-Resolution acquisition (SR) to the DAQ HAT: sr_filter (FIR decimator), DAQ_K_SR_MODE registry, settings/UI hooks and fast-path integration. Fixes USB waveform delivery and latency: per-batch flush limits (USB_WAVE_FLUSH_HZ), TX FIFO static_asserts, and transport coalescing (usb_stream_batch_begin/end) to eliminate dropped frames and reduce summary stalls. Adds per-stage cycle profiler (daq_perf), C6 Firmware screen + DDP v9 MB FWINFO/MB FW_APPLY tunnel, async snapshot/apply in api_core, version bumps (desktop/firmware), tests and UI/iOS support for SR and robustness fixes.

+++ Firmware/DAQ_HAT/ESP32P4/src/dsp/power_dsp.c
@@ -98,8 +98,14 @@ static void stat_finish(const running_stat_t *s, stat_result_t *out)
     out->min  = (s->min == INFINITY)  ? 0.0f : s->min;
     out->max  = (s->max == -INFINITY) ? 0.0f : s->max;
     out->mean = (float)mean;
-    out->rms  = (float)sqrt(var + mean * mean);
-    out->std  = (float)sqrt(var);
+    // Narrow to float BEFORE the sqrt, not after. The P4 FPU is single
+    // precision (see the header's note), so a double sqrt() is a soft-float
+    // call costing hundreds of cycles -- and its extra precision was being
+    // thrown away by the cast into these float fields anyway. The sums stay
+    // double, so only the final rounding differs, by at most 1 float ULP.
+    // Measured: 6 of these per 10 Hz summary dominated usb_stream_send_stats.
+    out->rms  = sqrtf((float)(var + mean * mean));
+    out->std  = sqrtf((float)var);
 }
 
 // -----------------------------------------------------------------------------
